Check for obsolete closed tasks without downloading artifacts
Koji doesn't give us a straightfoward method to check for any task
output.
In order to know whether there is something to download, we need to
find the descdendents of the original task, then look at the srpm build output.
If there's some, we're good. If there's none, we need to rebuild the
srpm.
This adds two new calls to koji, hence makes the script run last longer.
It's not really needed in the day-to-day operations, but after a
few-weeks-long pause the issue with trying to download non-existing srpms starts to appear.
Hence it's disabled by default, but available with a function flag.
If we ever add CLI options to the script, it could be propagated up,
ATM it'd be probably too much hassle.

+++ bconds.py
@@ -1,4 +1,5 @@
 import functools
+import json
 import pathlib
 import re
 import subprocess
@@ -143,11 +144,23 @@ def koji_status(koji_id):
 
 
 def koji_task_is_obsolete(koji_id):
-    command = ('koji', 'download-task', koji_id, '--arch=src', '--noprogress')
-    koji_output = run(*command).stdout.splitlines()
-    if 'No files for download found.' in koji_output:
-        return True
-    return False
+    """
+    Returns False if there exists an output of a 'rebuildRPM' task associated with
+    given koji_id, meaning the task was not garbage collected yet.
+    Return True if the output can't be reached or is returned from Koji empty.
+    """
+    command = ('koji', 'call', 'getTaskDescendents', koji_id, '--json-output')
+    koji_output = json.loads(run(*command).stdout)
+    desc_task_id = None
+    for descendent in koji_output[koji_id]:
+        if descendent['method'] == 'rebuildSRPM':
+            desc_task_id = descendent['id']
+    if desc_task_id:
+        command = ('koji', 'call', 'listTaskOutput', str(desc_task_id), '--json-output')
+        koji_output = json.loads(run(*command).stdout)
+        if koji_output:
+            return False
+    return True
 
 
 def handle_existing_srpm(repopath, *, was_updated):
@@ -160,7 +173,7 @@ def handle_existing_srpm(repopath, *, was_updated):
     return None
 
 
-def handle_existing_koji_id(repopath, *, was_updated):
+def handle_existing_koji_id(repopath, *, was_updated, check_if_garbage_collected=False):
     koji_id_path = repopath / KOJI_ID_FILENAME
     if koji_id_path.exists():
         if was_updated:
@@ -174,10 +187,12 @@ def handle_existing_koji_id(repopath, *, was_updated):
                     f'removing {KOJI_ID_FILENAME}.')
                 koji_id_path.unlink()
                 return None
-            elif status == 'closed' and koji_task_is_obsolete(koji_task_id):
-                log(f'   • Koji task {koji_task_id} is long closed '
-                    f'and there\'s nothing to download; removing {KOJI_ID_FILENAME}.')
-                koji_id_path.unlink()
+            if check_if_garbage_collected:
+                if status == 'closed' and koji_task_is_obsolete(koji_task_id):
+                    log(f'   • Koji task {koji_task_id} is long closed '
+                        f'and there\'s nothing to download; removing {KOJI_ID_FILENAME}.')
+                    koji_id_path.unlink()
+                    return None
             else:
                 log(f'   • Koji task {koji_task_id} is {status}; '
                     f'not rebuilding (rm {KOJI_ID_FILENAME} to force).')
@@ -209,7 +224,8 @@ def scratchbuild_patched_if_needed(component_name, bcond_config, *, branch='', t
     if srpm := handle_existing_srpm(repopath, was_updated=news):
         bcond_config['srpm'] = srpm
         return False
-
+    # set check_if_garbage_collected=True if you're running the script after a long pause
+    # builds may have been successful but the srpms are long deleted - we need to submit them again
     if koji_id := handle_existing_koji_id(repopath, was_updated=news):
         bcond_config['koji_task_id'] = koji_id
         return False
